+++ lib/fluent_ui.dart
@@ -18,7 +18,8 @@ export 'src/controls/inputs/drop_down_button.dart';
 export 'src/controls/inputs/icon_button.dart';
 export 'src/controls/inputs/radio_button.dart';
 export 'src/controls/inputs/split_button.dart';
-export 'src/controls/inputs/toggle.dart';
+export 'src/controls/inputs/toggle_button.dart';
+export 'src/controls/inputs/toggle_switch.dart';
 
 export 'src/controls/items/list_cell.dart';